feat: add workflows for Homebrew and NPM releases, implement draft release job, and refactor CI workflow generation
