+++ crates/forge_ci/Cargo.toml
@@ -6,10 +6,8 @@ edition = "2021"
 [dependencies]
 serde.workspace = true
 serde_json.workspace = true
+gh-workflow-tailcall.workspace = true
+indexmap.workspace = true
 
 [dev-dependencies]
-gh-workflow-tailcall.workspace = true
 pretty_assertions.workspace = true
-serde_json.workspace = true
-serde.workspace = true
-indexmap.workspace = true
