+++ crates/forge_ci/src/jobs/mod.rs
@@ -1,0 +1,7 @@
+//! Jobs for CI workflows
+
+pub mod build;
+pub mod draft_release;
+pub mod homebrew;
+pub mod npm;
+pub mod release_drafter;
