+++ crates/forge_ci/src/lib.rs
@@ -1,1 +1,3 @@
+pub mod jobs;
 pub mod matrix;
+pub mod workflow;
